update to rust-crypto and rename rcw to cryptoxide
* fix some warnings
* remove some comments
* improve documentation
