+++ cardano/Cargo.toml
@@ -11,7 +11,7 @@ keywords = [ "Cardano", "Wallet", "Crypto" ]
 [dependencies]
 serde = "1.0"
 serde_derive = "1.0"
-rcw = { path = "../rcw" }
+cryptoxide = { path = "../cryptoxide" }
 cbor_event = { path = "../cbor_event" }
 
 [dev-dependencies]
