+++ cryptoxide/src/digest.rs
