+++ cryptoxide/src/hkdf.rs
@@ -5,7 +5,7 @@
 // except according to those terms.
 
 //! This module implements the HMAC-based Extract-and-Expand Key
-//! Derivation Function as specified by  https://tools.ietf.org/html/rfc5869.
+//! Derivation Function as specified by <https://tools.ietf.org/html/rfc5869>.
 
 use std::iter::repeat;
 use cryptoutil::copy_memory;
@@ -21,7 +21,7 @@ use mac::Mac;
 /// * digest - The digest function to use.
 /// * salt - The optional salt value (a non-secret random value) to use.
 /// * ikm - The input keying material to use.
-/// * prk - The output buffer to fill with a digest.output_bytes() length
+/// * prk - The output buffer to fill with a `digest.output_bytes()` length
 ///         pseudo random key.
 pub fn hkdf_extract<D: Digest>(mut digest: D, salt: &[u8], ikm: &[u8], prk: &mut [u8]) {
     assert!(prk.len() == digest.output_bytes());
@@ -38,7 +38,7 @@ pub fn hkdf_extract<D: Digest>(mut digest: D, salt: &[u8], ikm: &[u8], prk: &mut
 ///
 /// # Arguments
 /// * digest - The digest function to use.
-/// * prk - The pseudorandom key of at least digest.output_bytes() octets.
+/// * prk - The pseudorandom key of at least `digest.output_bytes()` octets.
 /// * info - The optional context and application specific information to use.
 /// * okm - The output buffer to fill with the derived key value.
 pub fn hkdf_expand<D: Digest>(mut digest: D, prk: &[u8], info: &[u8], okm: &mut [u8]) {
