+++ cryptoxide/src/simd.rs
