+++ cryptoxide/src/symmetriccipher.rs
