+++ cryptoxide/src/util.rs
