--- rcw/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "rcw"
-version = "0.2.36"
-authors = ["The Rust-Crypto Project Developers"]
-license = "MIT/Apache-2.0"
-homepage = "https://github.com/DaGenix/rust-crypto/"
-repository = "https://github.com/DaGenix/rust-crypto/"
-description = "A (mostly) pure-Rust implementation of various common cryptographic algorithms."
-keywords = [ "Crypto", "MD5", "Sha1", "Sha2", "AES" ]
-
-[build-dependencies]
-
-[dependencies]
-
-#[lib]
-#crate-type = ["cdylib"]
