--- rcw/src/lib.rs
@@ -1,27 +1,0 @@
-// Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
-// http://www.apache.org/licenses/LICENSE-2.0> or the MIT license
-// <LICENSE-MIT or http://opensource.org/licenses/MIT>, at your
-// option. This file may not be copied, modified, or distributed
-// except according to those terms.
-
-pub mod blake2b;
-pub mod blake2s;
-pub mod buffer;
-pub mod aead;
-mod symmetriccipher;
-pub mod chacha20;
-pub mod chacha20poly1305;
-mod cryptoutil;
-pub mod curve25519;
-pub mod digest;
-pub mod ed25519;
-pub mod hmac;
-pub mod hkdf;
-pub mod mac;
-pub mod pbkdf2;
-pub mod poly1305;
-pub mod sha2;
-pub mod sha3;
-mod simd;
-mod step_by;
-pub mod util;
