+++ storage/src/pack.rs
@@ -16,8 +16,8 @@ use std::io::SeekFrom;
 use std::io;
 use std::io::{Write,Read,Seek,ErrorKind};
 use std::fs;
-use rcw::blake2b;
-use rcw::digest::Digest;
+use cryptoxide::blake2b;
+use cryptoxide::digest::Digest;
 use types::HASH_SIZE;
 use bloom;
 use types::BlockHash;
